requirements.txt
# FastAPI相关依赖
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# 原有后端依赖
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.26
PyPDF2>=3.0.1
pypdf>=3.17.0
python-docx>=1.1.0
chromadb>=0.4.18
scikit-learn>=1.3.2
matplotlib>=3.8.2
plotly>=5.17.0
Pillow>=10.1.0

# 多模态分析依赖
opencv-python>=4.8.1.78
mediapipe>=0.10.8
librosa>=0.10.1
pyaudio>=0.2.11
speech-recognition>=3.10.0
transformers>=4.36.0
torch>=2.1.1
torchvision>=0.16.1
soundfile>=0.12.1

# 其他工具依赖
requests>=2.31.0
aiofiles>=23.2.1
jinja2>=3.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2

# 数据持久化依赖
aiosqlite>=0.19.0
aioredis>=2.0.1
redis>=5.0.1